Moving taskB model to models.py.
